+++ src/interfaces/format-name-props.interface.ts
@@ -1,0 +1,4 @@
+export interface FormatNameProps{
+  chosenNameFormat: 'kebabCase' | 'pascalCase';
+  nameComponent: string;
+}