+++ src/utils/write-file.util.ts
@@ -1,0 +1,11 @@
+type WriteFilePros = {
+  pathFile: string;
+  contentFile: string;
+}
+
+export const writeFile = ({pathFile, contentFile}: WriteFilePros) => {
+  const fs = require('fs');
+  if(!fs.existsSync(pathFile)){
+    fs.writeFileSync(pathFile, contentFile);
+  }
+}