+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Change Log
+
+## [0.1.2] - 2022-07-24
+### Fixed
+  - Fix a bug of error import styles.
+    - Issue: [#15](https://github.com/viniciusanchieta/generate-component-and-style/issues/15)
+### Added
+  - Add a new feature to choose file name format.
+    - Two new options: KebabCase and PascalCase.
+
 ## [0.0.8] - 2022-07-14
 ### Fixed
 - Fixed a bug of special characters in the function.